+++ .vscode/settings.json
@@ -8,5 +8,6 @@
     "**/.DS_Store": true,
     "**/Thumbs.db": true,
     "**/node_modules": true
-  }
+  },
+  "typescript.tsdk": "node_modules\\typescript\\lib"
 }
