--- core/import.ts
@@ -1,7 +1,0 @@
-import { default as Axios } from 'axios'
-import { default as NextLink } from 'next/link'
-import { default as Router } from 'next/router'
-
-export const axios = Axios
-export const Link = NextLink
-export const router = Router
