+++ package.json
@@ -16,6 +16,10 @@
     "*.{ts,tsx}": "yarn pre-commit"
   },
   "dependencies": {
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@fontsource/inter": "^4.5.11",
+    "@mui/material": "^5.9.1",
     "@reduxjs/toolkit": "^1.8.3",
     "axios": "^0.27.2",
     "cookie": "^0.5.0",
