+++ src/core/firebase/client.ts
