+++ src/core/index.ts
@@ -1,0 +1,9 @@
+import { default as ReduxStore } from '@core/config/store'
+import { default as Axios } from 'axios'
+import { default as NextLink } from 'next/link'
+import { default as Router } from 'next/router'
+
+export const axios = Axios
+export const Link = NextLink
+export const router = Router
+export const store = ReduxStore
