+++ src/pages/_app.tsx
@@ -1,0 +1,22 @@
+import type { AppProps } from 'next/app'
+
+import '@fontsource/inter/400.css'
+import '@fontsource/inter/500.css'
+import '@fontsource/inter/600.css'
+import '@fontsource/inter/700.css'
+
+import { AuthContextProvider } from '@context/Auth'
+import { store } from '@core'
+import { Provider } from 'react-redux'
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return (
+    <Provider store={store}>
+      <AuthContextProvider>
+        <Component {...pageProps} />
+      </AuthContextProvider>
+    </Provider>
+  )
+}
+
+export default MyApp
