+++ src/pages/_document.tsx
@@ -1,0 +1,15 @@
+import { Head, Html, Main, NextScript } from 'next/document'
+
+const Document = () => {
+  return (
+    <Html lang='en'>
+      <Head />
+      <body>
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  )
+}
+
+export default Document
