+++ src/tools/typography.ts
@@ -1,0 +1,12 @@
+export const typography = {
+  body: {
+    fontSize: 17,
+    lineHeight: 1.475,
+    letterSpacing: '-0.013em',
+  },
+  caption: {
+    fontSize: 14,
+    lineHeight: 1.2857,
+    letterSpacing: '0',
+  },
+}
