+++ src/types/http.ts
@@ -1,0 +1,4 @@
+export interface FetchRequest<C = {}, M = {}> {
+  response: C
+  payload: M
+}
