+++ src/types/page.ts
@@ -1,0 +1,4 @@
+import type { User } from 'firebase/auth'
+import type { NextPage } from 'next'
+
+export type NextProtectedPage<T = {}> = NextPage<{ user: User | null } & T>
