+++ tsconfig.json
@@ -14,11 +14,26 @@
     "isolatedModules": true,
     "jsx": "preserve",
     "incremental": true,
-    "baseUrl": ".",
+    "baseUrl": "./src",
     "paths": {
-      "*": ["*"]
+      "@core": ["core"],
+      "@core/*": ["core/*"],
+      "@pages": ["pages"],
+      "@pages/*": ["pages/*"],
+      "@store": ["store"],
+      "@store/*": ["store/*"],
+      "@tools": ["tools"],
+      "@tools/*": ["tools/*"],
+      "@context": ["context"],
+      "@context/*": ["context/*"],
+      "@components": ["components"],
+      "@components/*": ["components/*"],
+      "@server": ["core/server"],
+      "@server/*": ["core/server/*"],
+      "@type": ["types"],
+      "@type/*": ["types/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["global.d.ts", "next-env.d.ts", "**/*.ts", "**/*.tsx"],
   "exclude": ["node_modules"]
 }
